+++ checklist-dashboard/widget/widget-frontend/Checklist/package.json
@@ -1,0 +1,33 @@
+{
+    "name": "checklist-widget",
+    "version": "1.0.0",
+    "private": true,
+    "dependencies": {
+        "@material-ui/core": "^3.9.2",
+        "@material-ui/styles": "^3.0.0-alpha.10",
+        "@wso2-dashboards/widget": "^1.4.0",
+        "append-query": "^2.0.1",
+        "axios": "^0.18.0",
+        "axois": "^1.0.0",
+        "css-loader": "^2.1.0",
+        "react": "^16.1.1",
+        "react-dom": "^16.1.1",
+        "style-loader": "^0.23.1"
+    },
+    "scripts": {
+        "build": "webpack -p",
+        "build-dev": "webpack -d --config webpack.config.js --watch --progress"
+    },
+    "devDependencies": {
+        "webpack": "^3.5.6",
+        "ajv": "^5.2.2",
+        "babel-core": "^6.25.0",
+        "babel-loader": "^7.1.1",
+        "babel-preset-es2015": "^6.24.1",
+        "babel-preset-react": "^6.24.1",
+        "babel-register": "^6.26.0",
+        "copy-webpack-plugin": "^4.2.0",
+        "node-sass": "^4.5.3",
+        "sass-loader": "^6.0.6"
+    }
+}
