--- checklist-dashboard/widget/widget-frontend/hello-world/src/resources/widgetConf.json
@@ -1,6 +1,0 @@
-{
-    "name": "HelloWorld",
-    "id": "HelloWorld",
-    "thumbnailURL": "",
-    "configs": {}
-}
