+++ checklist-dashboard/widget/widget-frontend/package.json
@@ -1,0 +1,24 @@
+{
+    "name": "checklist-widget",
+    "version": "1.0.0",
+    "private": true,
+    "dependencies": {
+        "react": "^16.1.1",
+        "react-dom": "^16.1.1"
+    },
+    "scripts": {
+        "build": "webpack -p"
+    },
+    "devDependencies": {
+        "webpack": "^3.5.6",
+        "ajv": "^5.2.2",
+        "babel-core": "^6.25.0",
+        "babel-loader": "^7.1.1",
+        "babel-preset-es2015": "^6.24.1",
+        "babel-preset-react": "^6.24.1",
+        "babel-register": "^6.26.0",
+        "copy-webpack-plugin": "^4.2.0",
+        "node-sass": "^4.5.3",
+        "sass-loader": "^6.0.6"
+    }
+}
