+++ checklist-dashboard/widget/widget-frontend/src/resources/widgetConf.json
@@ -1,0 +1,6 @@
+{
+    "name": "Checklist",
+    "id": "Checklist",
+    "thumbnailURL": "",
+    "configs": {}
+}
